Former-commit-id: 2c75d7bdfb9919497ecc9564a050042ccc9fd248

+++ hq-server/src/main/java/org/hyperic/hq/appdef/server/session/AgentManagerImpl.java
@@ -1243,7 +1243,7 @@ public class AgentManagerImpl implements AgentManager, ApplicationContextAware {
     
 // XXX needs javadoc!
     @Transactional(readOnly=true)
-    public int getNumAutoUpdatingAgents() {
+    public long getNumAutoUpdatingAgents() {
         return agentPluginStatusDAO.getNumAutoUpdatingAgents();
     }
